+++ packages/cubejs-backend-native/src/transport.rs
@@ -6,7 +6,7 @@ use cubeclient::models::{V1Error, V1LoadRequestQuery, V1LoadResponse, V1MetaResp
 use cubesql::{
     di_service,
     sql::AuthContextRef,
-    transport::{CubeReadStream, LoadRequestMeta, MetaContext, TransportService},
+    transport::{CubeStreamReceiver, LoadRequestMeta, MetaContext, TransportService},
     CubeError,
 };
 use serde_derive::Serialize;
@@ -164,37 +164,48 @@ impl TransportService for NodeBridgeTransport {
         }
     }
 
-    fn load_stream(
+    async fn load_stream(
         &self,
         query: V1LoadRequestQuery,
         ctx: AuthContextRef,
         meta: LoadRequestMeta,
-    ) -> Result<Arc<dyn CubeReadStream>, CubeError> {
+    ) -> Result<CubeStreamReceiver, CubeError> {
         trace!("[transport] Request ->");
 
-        let native_auth = ctx
-            .as_any()
-            .downcast_ref::<NativeAuthContext>()
-            .expect("Unable to cast AuthContext to NativeAuthContext");
+        loop {
+            let native_auth = ctx
+                .as_any()
+                .downcast_ref::<NativeAuthContext>()
+                .expect("Unable to cast AuthContext to NativeAuthContext");
 
-        let request_id = Uuid::new_v4().to_string();
-        let extra = serde_json::to_string(&LoadRequest {
-            request: TransportRequest {
-                id: format!("{}-span-{}", request_id, 1),
-                meta: Some(meta),
-            },
-            query,
-            session: SessionContext {
-                user: native_auth.user.clone(),
-                superuser: native_auth.superuser,
-            },
-        })?;
+            let request_id = Uuid::new_v4().to_string();
+            let extra = serde_json::to_string(&LoadRequest {
+                request: TransportRequest {
+                    id: format!("{}-span-{}", request_id, 1),
+                    meta: Some(meta.clone()),
+                },
+                query: query.clone(),
+                session: SessionContext {
+                    user: native_auth.user.clone(),
+                    superuser: native_auth.superuser,
+                },
+            })?;
 
-        call_js_with_stream_as_callback(
-            self.channel.clone(),
-            self.on_load_stream.clone(),
-            Some(extra),
-        )
+            let res = call_js_with_stream_as_callback(
+                self.channel.clone(),
+                self.on_load_stream.clone(),
+                Some(extra),
+            )
+            .await;
+
+            if let Err(e) = &res {
+                if e.message.to_lowercase() == "continue wait" {
+                    continue;
+                }
+            }
+
+            break res;
+        }
     }
 }