+++ packages/cubejs-hive-driver/driver/HiveDriver.js
@@ -62,7 +62,9 @@ class HiveDriver extends BaseDriver {
    * Class constructor.
    */
   constructor(config = {}) {
-    super();
+    super({
+      testConnectionTimeout: config.testConnectionTimeout,
+    });
 
     const dataSource =
       config.dataSource ||