+++ packages/cubejs-oracle-driver/driver/OracleDriver.js
@@ -58,7 +58,9 @@ class OracleDriver extends BaseDriver {
    * Class constructor.
    */
   constructor(config = {}) {
-    super();
+    super({
+      testConnectionTimeout: config.testConnectionTimeout,
+    });
 
     const dataSource =
       config.dataSource ||