+++ packages/cubejs-query-orchestrator/driver/utils.js
@@ -1,8 +1,0 @@
-const { cancelCombinator } = require('../dist/src/driver/utils');
-
-process.emitWarning(
-  'Using absolute import with @cubejs-backend/query-orchestrator is deprecated',
-  'DeprecationWarning'
-);
-
-module.exports = cancelCombinator;