+++ packages/cubejs-query-orchestrator/orchestrator/ContinueWaitError.js
@@ -1,8 +1,0 @@
-const { ContinueWaitError } = require('../dist/src/orchestrator/ContinueWaitError');
-
-process.emitWarning(
-  'Using absolute import with @cubejs-backend/query-orchestrator is deprecated',
-  'DeprecationWarning'
-);
-
-module.exports = ContinueWaitError;