+++ packages/cubejs-schema-compiler/adapter/BaseMeasure.js
@@ -1,8 +1,0 @@
-const { BaseMeasure } = require('../dist/src/adapter/BaseMeasure');
-
-process.emitWarning(
-  'Using absolute import with @cubejs-backend/schema-compiler is deprecated',
-  'DeprecationWarning'
-);
-
-module.exports = BaseMeasure;